Update repo_scripts/setup_env.sh
nice

+++ repo_scripts/setup_env.sh
@@ -2,7 +2,10 @@
 
 # change directory to the project's root, one-level up from the current directory
 cd $(dirname "$0")/..
-
+if [[ -f .env.server.local ]] || [[ -f .env.shared.local ]]; then
+  echo "Error: local env files already exist."
+  exit 1
+fi
 # Copy contents of .env.*.example files to new .env.*.local files
 
 cat .env.server.example > .env.server.local
